+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Addded
 
+- `/txs/{hash}/utxos`
+  - `consumed_by_tx` field
 - `/epochs/{number}/parameters` and `/epochs/latest/parameters`
   - `cost_models_raw` field, list variant of cost_models without name mapping